+++ App.tsx
@@ -1,0 +1,115 @@
+import React from 'react';
+import {
+  Platform,
+  StatusBar,
+  StyleSheet,
+  View,
+  ActivityIndicator
+} from 'react-native';
+import { AppLoading } from 'expo';
+import * as Font from 'expo-font';
+import { Asset } from 'expo-asset';
+import { Ionicons } from '@expo/vector-icons';
+import ApiKeys from './constants/ApiKeys';
+import * as firebase from 'firebase';
+
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
+import { store, persistor } from './redux-persist/store';
+
+import AppNavigator from './navigation/AppNavigator';
+// import MainTabNavigator from "./navigation/MainTabNavigator";
+
+interface State {
+  status: 'loading' | 'authenticated' | 'not_authenticated';
+}
+
+
+export default class App extends React.Component {
+  /**
+   * The proper way to do this would just be to directly call your navigation
+   */
+  state: State = {
+    status: 'loading',
+  };
+
+  componentWillMount() {
+    // Initialize firebase...
+    if (!firebase.apps.length) {
+      firebase.initializeApp(ApiKeys.FirebaseConfig);
+    }
+
+    firebase.auth().onAuthStateChanged(this.onAuthStateChanged);
+  }
+
+  render() {
+    if (this.state.status === 'loading') {
+      return (
+        <AppLoading
+          startAsync={this.loadResourcesAsync as any}
+          onError={this.handleLoadingError}
+          onFinish={this.handleFinishLoading}
+        />
+      );
+    } else {
+      return (
+        <Provider store={store}>
+          <PersistGate persistor={persistor} loading={this.renderLoading()}>
+            <View style={styles.container}>
+              {Platform.OS === 'ios' && <StatusBar barStyle="default" />}
+              {/* {(this.state.status === 'authenticated') && <MainTabNavigator />} */}
+              {/* {(this.state.status === 'not_authenticated') &&  <AppNavigator />} */}
+              <AppNavigator />
+            </View>
+          </PersistGate>
+        </Provider>
+      );
+    }
+  }
+
+  private loadResourcesAsync = () => {
+    return Promise.all([
+      Asset.loadAsync([
+        require('./assets/images/robot-dev.png'),
+        require('./assets/images/robot-prod.png')
+      ]),
+      Font.loadAsync({
+        // This is the font that we are using for our tab bar
+        ...Ionicons.font,
+        'roboto-black': require('./assets/fonts/Roboto-Black.ttf'),
+        'roboto-regular': require('./assets/fonts/Roboto-Regular.ttf'),
+        'roboto-medium': require('./assets/fonts/Roboto-Medium.ttf'),
+        'roboto-bold': require('./assets/fonts/Roboto-Bold.ttf'),
+        'heebo-black': require('./assets/fonts/Heebo-Black.ttf'),
+        'heebo-thin': require('./assets/fonts/Heebo-Thin.ttf')
+      })
+    ]);
+  };
+
+  private handleLoadingError = error => {
+    // In this case, you might want to report the error to your error
+    // reporting service, for example Sentry
+    console.warn(error);
+  };
+
+  private handleFinishLoading = () => {
+    this.setState({ isLoadingComplete: true });
+  };
+
+  private onAuthStateChanged = user => {
+    this.setState({ isAuthenticated: !!user });
+  };
+
+  private renderLoading = () => (
+    <View style={styles.container}>
+      <ActivityIndicator size="large" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F4F1DE'
+  }
+});
