+++ package.json
@@ -57,9 +57,12 @@
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
+    "@types/expo": "32.0.0",
+    "@types/react-native": "0.57.0",
     "babel-preset-expo": "^5.1.0",
     "jest-expo": "^33.0.0",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "typescript": "^3.5.3"
   },
   "private": true
 }
