+++ screens/questionnaires/T1/Extra_2_Screen_T1.js
@@ -19,7 +19,6 @@ import RadioGroup, { Radio } from "react-native-radio-input";
 import { styles } from "./style";
 
 import * as firebase from "firebase";
-import { transform } from "@babel/core";
 
 export default class Extra_2_Screen_T1 extends React.Component {
   constructor(props) {
